==> experiments/MIMIC-ECG/7_create_pred_dataset.ipynb <==
import os
import pandas as pd
import numpy as np
import json
from tqdm import tqdm
from sklearn.model_selection import train_test_split
from sklearn.metrics import roc_auc_score
import re

### Load the dataset with the predictions

df = pd.read_csv('../../data/mimic-acute-mi_modelling_with_preds.csv')
#df[df['STEMI']==1].head()
display(df.head())
print(df.shape)

columns_to_transform = ['pred_st_elevation', 'pred_st_depression', 'pred_t_wave', 'pred_Q2', 'pred_troponin']

# Define the sigmoid function
def sigmoid(x):
    return 1 / (1 + np.exp(-x))

# Apply the sigmoid transformation to the specified columns
df[columns_to_transform] = df[columns_to_transform].apply(sigmoid)

# Display the dataframe to check the transformation
print(df[columns_to_transform].head())

## Check AUCs are ok
print(df.columns)
print(roc_auc_score(df['st_elevation'], df['pred_st_elevation']))
print(roc_auc_score(df['st_depression'], df['pred_st_depression']))
print(roc_auc_score(df['t_wave'], df['pred_t_wave']))
# print(roc_auc_score(df['Acute_MI'], df['pred_st_elevation']))
print(roc_auc_score(df['troponin'], df['pred_troponin']))
print(roc_auc_score(df['Q2'], df['pred_Q2']))

print(df.shape)
print(len(df['hadm_id'].unique()))

### Improve the medical note representation

print(df['simple_note'][5])

import re

# Your original string
text = "Hello,\nthis is a string with strange symbols! @#$%^&*()"

# Remove all non-alphanumeric characters and replace them with a space
cleaned_text = re.sub(r'[^a-zA-Z0-9\s]', '', text)

# Replace newline characters with a space
cleaned_text = cleaned_text.replace('\n', ' ')

# Remove extra spaces
cleaned_text = re.sub(r'\s+', ' ', cleaned_text).strip()

print(cleaned_text)

SECTIONS = ['Chief Complaint'] #, 'Past Medical History'
def clean_note(note):
    cleaned = ''
    for heading in SECTIONS:
        content = note.split(heading+':')
        if len(content) == 1: continue
        content = re.split(r'\n[^\n:]+:', content[1])[0] # find next section heading and cut
        cleaned += heading+':' + content + '\n'
    # Remove all non-alphanumeric characters and replace them with a space
    cleaned_text = re.sub(r'[^a-zA-Z0-9\s]', '', cleaned)

    # Replace newline characters with a space
    cleaned_text = cleaned_text.replace('\n', ' ')

    # Remove extra spaces
    cleaned_text = re.sub(r'\s+', ' ', cleaned_text).strip()
    return cleaned_text

df['chief_complaint'] = df['full_note'].map(clean_note)

print(df['chief_complaint'][2])

print(df['full_note'][1])

### Check on some conversations

with open('../../data/train_conversations_with_preds.json') as f:
    conversations = json.load(f)

# Function to filter and print the conversation by specific ID
def print_conversation_by_id(conversations, target_id):
    for item in conversations:
        if item.get('id') == target_id:
            print(f"Conversation with ID {target_id}:")
            print(json.dumps(item, indent=4))
            print("\n" + "-"*40 + "\n")
            return  # Exit after finding the conversation
    print(f"No conversation found with ID {target_id}")

# Example usage:
target_id = 43919386  # Replace this with the desired ID
print_conversation_by_id(conversations, target_id)
print(len(conversations))

with open('../../data/train_conversations_with_preds.json') as f:
    conversations = json.load(f)

N = 5  # Number of elements to display

for i, item in enumerate(conversations[-5:], start=1):
    print(f"Element {i}:")
    print(json.dumps(item, indent=4))
    print("\n" + "-"*40 + "\n")

with open('../../data/train_conversations_with_preds.json') as f:
    conversations = json.load(f)

N = 5  # Number of elements to display

for i, item in enumerate(conversations[-5:], start=1):
    print(f"Element {i}:")
    print(json.dumps(item, indent=4))
    print("\n" + "-"*40 + "\n")

### Split dataset into test and training dataframe

# Load the JSON file
with open('../../data/test_conversations.json', 'r') as file:
    data = json.load(file)

# Extract all 'id' values into a list
test_ids = [float(item['id']) for item in data]

# Display the list of ids
print(len(test_ids))

# Load the JSON file
with open('../../data/train_conversations.json', 'r') as file:
    data = json.load(file)

# Extract all 'id' values into a list
train_ids = [float(item['id']) for item in data]

# Display the list of ids
print(len(train_ids))

train_df = df[df['study_id'].isin(train_ids)]
print(train_df.shape)
test_df = df[df['study_id'].isin(test_ids)]
print(test_df.shape)

### Regenerate conversation files

with open('../../data/conv_templates_with_preds.json') as f: #diag_q
    conv_templates = json.load(f)

import random

def generate_conversations(df, conv_templates):
    conversations = []
    
    for idx, row in df.iterrows():
        idx_1 = row['study_id']
        conv = {
            "id": row['study_id'],
            "image": f"{idx_1}.jpeg",
            "conversations": []
        }

        if row['STEMI']:
            conv['diagnosis'] = 'STEMI'
        elif row['NSTEMI']:
            conv['diagnosis'] = 'NSTEMI'
        else:
            conv['diagnosis'] = 'NO ACS'

        conv['conversations'].append({
            "from": "human",
            "value": random.choice(conv_templates['st_elev_q']).replace('{INSERT PATIENT NOTE}', row['chief_complaint'])#.replace("{PREDICTIONS HERE}", f"ST elevation prediction: {round(row['pred_st_elevation'],2)}  ST depression prediction: {round(row['pred_st_depression'],2)} T-wave Inversion prediction {round(row['pred_t_wave'],2)}")
        })

        if row['st_elevation']:
            conv['conversations'].append({
                "from": "gpt",
                "value": random.choice(conv_templates['st_elev_a_true'])
            })
            conv['conversations'].append({
                "from": "human",
                "value": random.choice(conv_templates['st_depres_twi_q']).replace('Your prior assessment was: {INSERT PRIOR ANSWER}', '')#.replace("{PREDICTIONS HERE}", f"ST elevation prediction: {round(row['pred_st_elevation'],2)}  ST depression prediction: {round(row['pred_st_depression'],2)} T-wave Inversion prediction {round(row['pred_t_wave'],2)}")
            })
            if row['st_depression'] or row['t_wave']:
                conv['conversations'].append({
                    "from": "gpt",
                    "value": random.choice(conv_templates['st_depres_twi_a_true'])
                })
                conv['conversations'].append({
                    "from": "human",
                    "value": random.choice(conv_templates['troponin_q']).replace('Your prior assessment was: {INSERT PRIOR ANSWER}', '')#.replace("{PREDICTIONS HERE}", f"ST elevation prediction: {round(row['pred_st_elevation'],2)}  ST depression prediction: {round(row['pred_st_depression'],2)} T-wave Inversion prediction {round(row['pred_t_wave'],2)}")
                })
            else:
                conv['conversations'].append({
                    "from": "gpt",
                    "value": random.choice(conv_templates['st_depres_twi_a_false'])
                })
                conv['conversations'].append({
                    "from": "human",
                    "value": random.choice(conv_templates['troponin_q']).replace('Your prior assessment was: {INSERT PRIOR ANSWER}', '')#.replace("{PREDICTIONS HERE}", f"ST elevation prediction: {round(row['pred_st_elevation'],2)}  ST depression prediction: {round(row['pred_st_depression'],2)} T-wave Inversion prediction {round(row['pred_t_wave'],2)}")
                })
            
            if row['troponin']:
                conv['conversations'].append({
                    "from": "gpt",
                    "value": random.choice(conv_templates['troponin_a_true'])
                })
                conv['conversations'].append({
                    "from": "human",
                    "value": random.choice(conv_templates['diag_q_trop']).replace('Your prior assessment was: {INSERT PRIOR ANSWER}', '').replace('{INSERT_RESULT}', f"{row['valuenum']},{row['comments']}")#.replace("{PREDICTIONS HERE}", f"ST elevation prediction: {round(row['pred_st_elevation'],2)}  ST depression prediction: {round(row['pred_st_depression'],2)} T-wave Inversion prediction {round(row['pred_t_wave'],2)}")
                })
            else:
                conv['conversations'].append({
                    "from": "gpt",
                    "value": random.choice(conv_templates['troponin_a_false'])
                })
                conv['conversations'].append({
                    "from": "human",
                    "value": random.choice(conv_templates['diag_q']).replace('Your prior assessment was: {INSERT PRIOR ANSWER}', '')#.replace("{PREDICTIONS HERE}", f"ST elevation prediction: {round(row['pred_st_elevation'],2)}  ST depression prediction: {round(row['pred_st_depression'],2)} T-wave Inversion prediction {round(row['pred_t_wave'],2)}")
                })

            if row['STEMI']:
                conv['conversations'].append({
                    "from": "gpt",
                    "value": random.choice(conv_templates['diag_a_stemi'])
                })
            else:
                conv['conversations'].append({
                    "from": "gpt",
                    "value": random.choice(conv_templates['diag_a_normal'])
                })
        else:
            conv['conversations'].append({
                "from": "gpt",
                "value": random.choice(conv_templates['st_elev_a_false'])
            })
            conv['conversations'].append({
                "from": "human",
                "value": random.choice(conv_templates['st_depres_twi_q']).replace('Your prior assessment was: {INSERT PRIOR ANSWER}', '')#.replace("{PREDICTIONS HERE}", f"ST elevation prediction: {round(row['pred_st_elevation'],2)}  ST depression prediction: {round(row['pred_st_depression'],2)} T-wave Inversion prediction {round(row['pred_t_wave'],2)}")
            })
            if row['st_depression'] or row['t_wave']:
                conv['conversations'].append({
                    "from": "gpt",
                    "value": random.choice(conv_templates['st_depres_twi_a_true'])
                })
                conv['conversations'].append({
                    "from": "human",
                    "value": random.choice(conv_templates['troponin_q']).replace('Your prior assessment was: {INSERT PRIOR ANSWER}', '')#.replace("{PREDICTIONS HERE}", f"ST elevation prediction: {round(row['pred_st_elevation'],2)}  ST depression prediction: {round(row['pred_st_depression'],2)} T-wave Inversion prediction {round(row['pred_t_wave'],2)}")
                })
            else:
                conv['conversations'].append({
                    "from": "gpt",
                    "value": random.choice(conv_templates['st_depres_twi_a_false'])
                })
                conv['conversations'].append({
                    "from": "human",
                    "value": random.choice(conv_templates['troponin_q']).replace('Your prior assessment was: {INSERT PRIOR ANSWER}', '')#.replace("{PREDICTIONS HERE}", f"ST elevation prediction: {round(row['pred_st_elevation'],2)}  ST depression prediction: {round(row['pred_st_depression'],2)} T-wave Inversion prediction {round(row['pred_t_wave'],2)}")
                })

            if row['troponin']:
                conv['conversations'].append({
                    "from": "gpt",
                    "value": random.choice(conv_templates['troponin_a_true'])
                })
                conv['conversations'].append({
                    "from": "human",
                    "value": random.choice(conv_templates['diag_q_trop']).replace('Your prior assessment was: {INSERT PRIOR ANSWER}', '').replace('{INSERT_RESULT}', f"{row['valuenum']},{row['comments']}")#.replace("{PREDICTIONS HERE}", f"ST elevation prediction: {round(row['pred_st_elevation'],2)}  ST depression prediction: {round(row['pred_st_depression'],2)} T-wave Inversion prediction {round(row['pred_t_wave'],2)}")
                })
            else:
                conv['conversations'].append({
                    "from": "gpt",
                    "value": random.choice(conv_templates['troponin_a_false'])
                })
                conv['conversations'].append({
                    "from": "human",
                    "value": random.choice(conv_templates['diag_q']).replace('Your prior assessment was: {INSERT PRIOR ANSWER}', '')#.replace("{PREDICTIONS HERE}", f"ST elevation prediction: {round(row['pred_st_elevation'],2)}  ST depression prediction: {round(row['pred_st_depression'],2)} T-wave Inversion prediction {round(row['pred_t_wave'],2)}")
                })
                
            if row['NSTEMI']:
                conv['conversations'].append({
                    "from": "gpt",
                    "value": random.choice(conv_templates['diag_a_nstemi'])
                })
            else:
                conv['conversations'].append({
                    "from": "gpt",
                    "value": random.choice(conv_templates['diag_a_normal'])
                })
        
        conversations.append(conv)

    return conversations


# train_df = train_df[~train_df['simple_note'].isna()]
# train_df = train_df[(df['troponin'] & ~df['valuenum'].isnull() & ~df['comments'].isnull()) | (df['troponin'] == 0)]

train_conversations = generate_conversations(train_df, conv_templates)
test_conversations = generate_conversations(test_df, conv_templates)

with open('../../data/train_conversations_simple.json', 'w') as f: #_with_preds
    json.dump(train_conversations, f)

with open('../../data/test_conversations_simple.json', 'w') as f: #_with_preds
    json.dump(test_conversations, f)


### Generate corresponding Single Turn QA Datasets 


def transform_to_qa(test_conversations):
    qa_list = []  # List to hold all QA pairs

    # Iterate over each conversation item with a progress bar
    for item in tqdm(test_conversations, desc='Processing conversations'):
        convs = item.get('conversations', [])
        conv_id = item.get('id')
        image = item.get('image')
        diagnosis = item.get('diagnosis')

        # Initialize index
        i = 0
        while i < len(convs):
            # Check if the current turn is from 'human'
            if convs[i].get('from') == 'human':
                human_msg = convs[i].get('value', '').strip().replace('<image>', '')

                # Check if the next turn exists and is from 'gpt'
                if i + 1 < len(convs) and convs[i + 1].get('from') == 'gpt':
                    gpt_msg = convs[i + 1].get('value', '').strip()

                    # Append the QA pair to the list, maintaining 'from' and 'value'
                    qa_list.append({
                        'id': conv_id,
                        'image': image,
                        'conversations': [
                            {
                                'from': 'human',
                                'value': human_msg
                            },
                            {
                                'from': 'gpt',
                                'value': gpt_msg
                            }
                        ],
                        #'diagnosis': diagnosis
                    })

                    # Move to the next pair
                    i += 2
                else:
                    # If there's no corresponding 'gpt' response, skip to next
                    i += 1
            else:
                # If the current turn is not from 'human', skip to next
                i += 1
    return qa_list

qa_list = transform_to_qa(test_conversations)

print(f"Total QA pairs extracted: {len(qa_list)}")

with open('../../data/test_conversations_single_qa_simple.json', 'w') as f: #_with_preds
    json.dump(qa_list, f)

qa_list = transform_to_qa(train_conversations)
print(f"Total QA pairs extracted: {len(qa_list)}")

with open('../../data/train_conversations_single_qa_simple.json', 'w') as f: #_with_preds
    json.dump(qa_list, f)

def transform_to_q_and_a(test_conversations):
    q_list = []  # List to hold all QA pairs
    a_list = []

    # Iterate over each conversation item with a progress bar
    for item in tqdm(test_conversations, desc='Processing conversations'):
        convs = item.get('conversations', [])
        conv_id = item.get('id')
        image = item.get('image')
        diagnosis = item.get('diagnosis')

        # Initialize index
        i = 0
        while i < len(convs):
            # Check if the current turn is from 'human'
            if convs[i].get('from') == 'human':
                human_msg = convs[i].get('value', '').strip().replace('<image>', '')

                # Check if the next turn exists and is from 'gpt'
                if i + 1 < len(convs) and convs[i + 1].get('from') == 'gpt':
                    gpt_msg = convs[i + 1].get('value', '').strip()

                    # Append the QA pair to the list, maintaining 'from' and 'value'
                    q_list.append({
                        'question_id': conv_id,
                        'image': image,
                        'text':  human_msg,
                        'category': diagnosis,
                        'number': i//2                    
                        })

                    a_list.append({
                        'question_id': conv_id,
                        'image': image,
                        'text':  gpt_msg,
                        'category': diagnosis,
                        'number': i//2                     
                        })

                    # Move to the next pair
                    i += 2
                else:
                    # If there's no corresponding 'gpt' response, skip to next
                    i += 1
            else:
                # If the current turn is not from 'human', skip to next
                i += 1
    return q_list, a_list

q_list, a_list = transform_to_q_and_a(test_conversations)

print(f"Total QA pairs extracted: {len(q_list)}")

# Save q_list to a JSON file
with open('../../data/test_simple_no_pred_LLaVA_acs_q.json', 'w') as f:
    for item in q_list:
        json.dump(item, f)
        f.write('\n')

# Save a_list to a JSON file
with open('../../data/test_simple_no_pred_LLaVA_acs_a.json', 'w') as f:
    for item in a_list:
        json.dump(item, f)
        f.write('\n')

q_list, a_list = transform_to_q_and_a(train_conversations)

print(f"Total QA pairs extracted: {len(q_list)}")

# Save q_list to a JSON file
with open('../../data/train_simple_no_pred_LLaVA_acs_q.json', 'w') as f:
    for item in q_list:
        json.dump(item, f)
        f.write('\n')

# Save a_list to a JSON file
with open('../../data/train_simple_no_pred_LLaVA_acs_a.json', 'w') as f:
    for item in a_list:
        json.dump(item, f)
        f.write('\n')

i=2
print(i//2)
